Added server routes and queries
Added routes for /ids and /course which rturn a list of valid
ids for a certain college and all the information for a course
of a certain id at a certain college respectivly.

Also added the relevent quiries and inputs to create and acesss
the courses collection which stores the data used by mentioned routes.